Combine analyzer and code fix tests
